test(task-migrate-clean-batch): 迁 project_world_dir 6 用例到 gui_unit_test
420.5 Step 4。整文件迁移，纯投影数学，6/6 用 IM_UNUSED(ctx)，零 g_state。

- 新增 test/unit-correctness/gui/test_project_world_dir.cpp（6 个 TEST）
- 删除 test/gui/functional/test_project_world_dir.cpp + 注册声明/调用点
- 文件头一句指向 git-ignored 工作制品的注释改写为机制本身

核验：断言 IM_CHECK 31 → EXPECT_TRUE 31 一一对应；
gui_test 静态用例数 554→548；gui_unit_test 21→27 全绿。
420.2 踩过的 imgui_te_stubs 链接坑未复发（该 target 已常驻链接 stubs）。
