+++ test/CMakeLists.txt
@@ -221,7 +221,8 @@ if(BUILD_GUI)
     "${PROJ_TEST_DIR}/unit-correctness/gui/test_defaults_diff.cpp"
     "${PROJ_TEST_DIR}/unit-correctness/gui/test_smoke.cpp"
     "${PROJ_TEST_DIR}/unit-correctness/gui/test_crystal_renderer.cpp"
-    "${PROJ_TEST_DIR}/unit-correctness/gui/test_state_reconcile.cpp")
+    "${PROJ_TEST_DIR}/unit-correctness/gui/test_state_reconcile.cpp"
+    "${PROJ_TEST_DIR}/unit-correctness/gui/test_project_world_dir.cpp")
   target_include_directories(gui_unit_test
     PRIVATE ${PROJ_TEST_DIR})
   target_link_libraries(gui_unit_test
