+++ test/gui/test_gui_main.cpp
@@ -396,7 +396,6 @@ int main(int argc, char** argv) {
   RegisterFaceNumberOverlayTests(engine);
   RegisterAutoEvRegressionTests(engine);
   RegisterLinkedEntriesTests(engine);
-  RegisterProjectWorldDirTests(engine);
   RegisterHandednessGuardTests(engine);
   RegisterLifecycleTests(engine);
   RegisterCompositePreviewTests(engine);
